Auth.py created

+++ crud.py
@@ -1,25 +1,9 @@
-import hashlib
 from fastapi import HTTPException
 from typing import List, Optional
+from auth import hash_password
 from models import Task, NewTask, User, NewUser, UpdateTask
 from database import cur, conn
 from datetime import datetime
-
-def authorithation(id: int, user_id: int):
-    cur.execute("SELECT user_id FROM tasks WHERE id = ?", (id,))
-    user = cur.fetchone()
-    if not user:
-        raise HTTPException(status_code=404, detail='Wrong id')
-    if user_id != user[0]:
-        raise HTTPException(status_code=403, detail='Not this user')
-
-def hashing(password: str) -> str:
-    return hashlib.sha256(password.encode()).hexdigest()
-
-def check_user(user_id: int):
-    cur.execute("SELECT id FROM users WHERE id = ?", (user_id, ))
-    if not cur.fetchone():
-        raise HTTPException(status_code=404, detail="User not found")
 
 def fetch_tasks() -> List[Task]:
     tasks = cur.fetchall()  # 3
@@ -28,14 +12,13 @@ def fetch_tasks() -> List[Task]:
 def fetch_task(task) -> Task:
     return Task(id=task[0], title=task[1], done=bool(task[2]), user_id=task[3], created=task[4])
 
-def get_tasks(user_id: Optional[int] = None, done: Optional[bool] = None, sort: str = None, limit: Optional[int] = None, title: Optional[str] = None) -> List[Task]:
+def get_tasks(current_user: User, done: Optional[bool] = None, sort: str = None, limit: Optional[int] = None, title: Optional[str] = None) -> List[Task]:
     query = "SELECT * FROM tasks" #1
     filters = []
     params = []
     order = None
-    if user_id is not None:
-        filters.append("user_id = ?")
-        params.append(user_id)
+    filters.append("user_id = ?")
+    params.append(current_user.id)
     if done is not None:
         filters.append("done = ?")
         params.append(int(done))
@@ -56,23 +39,22 @@ def get_tasks(user_id: Optional[int] = None, done: Optional[bool] = None, sort: 
     cur.execute(query, tuple(params))
     return fetch_tasks()
 
-def get_task_id(id: int) -> Task:
-    cur.execute("SELECT * FROM tasks WHERE id = ?", (id, ))
+def get_task_id(id: int, current_user: User) -> Task:
+    cur.execute("SELECT * FROM tasks WHERE id = ? AND user_id = ?", (id, current_user.id))
     task = cur.fetchone()
     if not task:
         raise HTTPException(status_code=404, detail='Nothing found with this id')
     return fetch_task(task)
 
-def add_task(newtask: NewTask) -> Task:
-    check_user(newtask.user_id)
-    cur.execute("INSERT INTO tasks (title, done, user_id, created_at) VALUES (?, ?, ?, ?)", (newtask.title, 0, newtask.user_id, datetime.now()))
+def add_task(newtask: NewTask, current_user: User) -> Task:
+    cur.execute("INSERT INTO tasks (title, done, user_id, created_at) VALUES (?, ?, ?, ?)", (newtask.title, 0, current_user.id, datetime.now()))
     conn.commit()
-    new_id = cur.lastrowid
-    return Task(id=new_id, title=newtask.title, done=False, user_id=newtask.user_id, created=str(datetime.now()))
 
-def change_task(id: int, user_id: int, newtask: UpdateTask) -> Task:
-    authorithation(id, user_id)
-    cur.execute("UPDATE tasks SET title = ? WHERE id = ? AND user_id = ?", (newtask.title, id, user_id))
+    new_id = cur.lastrowid
+    return Task(id=new_id, title=newtask.title, done=False, user_id=current_user.id, created=str(datetime.now()))
+
+def change_task(id: int,  newtask: UpdateTask, current_user: User) -> Task:
+    cur.execute("UPDATE tasks SET title = ? WHERE id = ? and user_id = ?", (newtask.title, id, current_user.id))
     conn.commit()
     cur.execute("SELECT * FROM tasks WHERE id = ?", (id, ))
     t = cur.fetchone()
@@ -80,13 +62,12 @@ def change_task(id: int, user_id: int, newtask: UpdateTask) -> Task:
         raise HTTPException(status_code=404, detail='Nothing found with this id')
     return fetch_task(t)
 
-def delete_task(id: int, user_id: int) -> List[Task]:
-    authorithation(id, user_id)
-    cur.execute("DELETE FROM tasks WHERE id = ?", (id, ))
+def delete_task(id: int, current_user: User) -> List[Task]:
+    cur.execute("DELETE FROM tasks WHERE id = ? AND user_id = ?", (id, current_user.id))
     conn.commit()
     if cur.rowcount == 0:
         raise HTTPException(status_code=404, detail='Task not found')
-    cur.execute("SELECT * FROM tasks")
+    cur.execute("SELECT * FROM tasks WHERE user_id = ?", (current_user.id, ))
     return fetch_tasks()
 
 def get_users() -> List[User]:
@@ -105,38 +86,29 @@ def add_user(newuser: NewUser) -> User:
     cur.execute("SELECT name FROM users WHERE name = ?", (newuser.username, ))
     if cur.fetchone():
         raise HTTPException(status_code=400, detail='There is a user with this username')
-    cur.execute("INSERT INTO users (name, password, created_at) VALUES (?, ?, ?)", (newuser.username, hashing(newuser.password), datetime.now()))
+    cur.execute("INSERT INTO users (name, hashed_password, created_at) VALUES (?, ?, ?)", (newuser.username, hash_password(newuser.password), datetime.now()))
     conn.commit()
     new_id = cur.lastrowid
     return User(id=new_id, username=newuser.username, created=str(datetime.now()))
 
-def get_user_task(user_id: int, done: Optional[bool] = None) -> List[Task]:
-    cur.execute("SELECT id FROM users WHERE id = ?", (user_id, ))
-    if not cur.fetchone():
-        raise HTTPException(status_code=404, detail='User not found')
+def get_user_task(current_user: User, done: Optional[bool] = None) -> List[Task]:
     if done is None:
-        cur.execute("SELECT * FROM tasks WHERE user_id = ?", (user_id, ))
+        cur.execute("SELECT * FROM tasks WHERE user_id = ?", (current_user.id, ))
     else:
-        cur.execute("SELECT * FROM tasks WHERE user_id = ? AND done = ?", (user_id, int(done)))
+        cur.execute("SELECT * FROM tasks WHERE user_id = ? AND done = ?", (current_user.id, int(done)))
     return fetch_tasks()
 
-def change_done(id: int, user_id: int) -> Task:
-    authorithation(id, user_id)
-    cur.execute("SELECT done FROM tasks WHERE id = ?", (id, ))
-    user = cur.fetchone()[0]
-    new_done = 1 if user == 0 else 0
-    cur.execute("UPDATE tasks SET done = ? WHERE id = ?", (new_done, id))
-    conn.commit()
-    cur.execute("SELECT * FROM tasks WHERE id = ?", (id,))
+def change_done(id: int, current_user: User) -> Task:
+    cur.execute("SELECT done FROM tasks WHERE id = ? AND user_id = ?", (id, current_user.id))
     task = cur.fetchone()
-    return fetch_task(task)
-
-def login(newuser: NewUser) -> int:
-    cur.execute("SELECT id FROM users WHERE name = ? AND password = ?", (newuser.username, hashing(newuser.password)))
-    user = cur.fetchone()
-    if not user:
-        raise HTTPException(status_code=404, detail='User not exists')
-    return user[0]
+    if not task:
+        raise HTTPException(status_code=404, detail='Task not found')
+    new_done = 1 if task[0] == 0 else 0
+    cur.execute("UPDATE tasks SET done = ? WHERE id = ? AND user_id = ?", (new_done, id, current_user.id))
+    conn.commit()
+    cur.execute("SELECT * FROM tasks WHERE id = ? AND user_id = ?", (id, current_user.id))
+    changed_task = cur.fetchone()
+    return fetch_task(changed_task)
 
 def get_user_by_username(username) -> User:
     cur.execute("SELECT id, name, created_at FROM users WHERE name = ?", (username, ))
@@ -144,5 +116,17 @@ def get_user_by_username(username) -> User:
     if not user:
         raise HTTPException(status_code=404, detail='User not found')
     return User(id=user[0], username=user[1], created=user[2])
+
+def profile(current_user: User):
+    return current_user
+
+def my_tasks(current_user: User) -> List[Task]:
+    cur.execute("SELECT * FROM tasks WHERE user_id = ?", (current_user.id, ))
+    return fetch_tasks()
+
+def completed_tasks(current_user: User) -> List[Task]:
+    cur.execute("SELECT * FROM tasks WHERE user_id = ? AND done = ?", (current_user.id, 1))
+    return fetch_tasks()
+
